handle exceptions

+++ src/main/java/com/digitalhouse/clinicaOdontologica/service/impl/OdontologoService.java
@@ -2,6 +2,7 @@ package com.digitalhouse.clinicaOdontologica.service.impl;
 
 
 import com.digitalhouse.clinicaOdontologica.entity.Odontologo;
+import com.digitalhouse.clinicaOdontologica.exception.ResourceNotFoundException;
 import com.digitalhouse.clinicaOdontologica.repository.IOdontologoRepository;
 import com.digitalhouse.clinicaOdontologica.service.IOdontologoService;
 import org.slf4j.Logger;
@@ -29,6 +30,8 @@ public class OdontologoService implements IOdontologoService {
             LOGGER.info("Odontologo guardado " + odontologoDB.toString());
         }else{
             LOGGER.info("No se guardo ningun odontologo");
+            throw new ResourceNotFoundException("El odontologo no fue guardado");
+
         }
 
         return odontologoDB ;
@@ -42,6 +45,8 @@ public class OdontologoService implements IOdontologoService {
             LOGGER.info("Odontologo encontrado " + odontologoDB.get().toString());
         }else{
             LOGGER.info("No se encontro ningun odontologo ");
+            throw new ResourceNotFoundException("El odontologo con id " + id +" no fue encontrado");
+
         }
         return odontologoDB ;
     }
@@ -54,6 +59,8 @@ public class OdontologoService implements IOdontologoService {
             LOGGER.info("Odontologos encontrados " + odontologosDB);
         }else{
             LOGGER.info("No se encontro ningun odontologo");
+            throw new ResourceNotFoundException("No se encontro ningun odontologo");
+
         }
 
         return odontologosDB ;
@@ -61,14 +68,28 @@ public class OdontologoService implements IOdontologoService {
 
     @Override
     public void updateOdontologo(Odontologo odontologo) {
-        odontologoRepository.save(odontologo);
-        LOGGER.info("Odontologo fue actualizado");
+        Optional<Odontologo> odontologoDB = odontologoRepository.findById(odontologo.getId());
+        if (odontologoDB.isPresent()){
+            odontologoRepository.save(odontologo);
+            LOGGER.info("Odontologo fue actualizado");
+        }else {
+            LOGGER.info("No se encontro el odontologo para actualizar");
+            throw new ResourceNotFoundException("El odontologo con id " + odontologo.getId() +" no fue encontrado para actualizar");
+        }
+
     }
 
     @Override
     public void deleteOdontologo(Integer id) {
-        odontologoRepository.deleteById(id);
-        LOGGER.info("Odontologo fue eliminado");
+        Optional<Odontologo> odontologoDB = odontologoRepository.findById(id);
+        if (odontologoDB.isPresent()){
+            odontologoRepository.deleteById(id);
+            LOGGER.info("Odontologo fue eliminado");
+        }else {
+            LOGGER.info("No se encontro el odontologo para eliminar");
+            throw new ResourceNotFoundException("El odontologo con id " + id +" no fue encontrado para eliminar");
+        }
+
     }
 
     @Override
@@ -79,6 +100,7 @@ public class OdontologoService implements IOdontologoService {
             LOGGER.info("Odontologo encontrado" + odontologoDB.toString());
         }else{
             LOGGER.info("No se encontro ningun odontologo ");
+            throw new ResourceNotFoundException("El odontologo con matriucla " + numeroMatricula +" no fue encontrado");
         }
         return odontologoDB ;
     }
@@ -91,6 +113,8 @@ public class OdontologoService implements IOdontologoService {
             LOGGER.info("Se encontraron odontologos");
         }else{
             LOGGER.info("No se encontro ningun odontologo ");
+            throw new ResourceNotFoundException("El odontologo con apellido " + apellido +" no fue encontrado");
+
         }
         return odontologosDB ;
     }
