+++ src/main/java/com/digitalhouse/clinicaOdontologica/service/impl/PacienteService.java
@@ -2,8 +2,11 @@ package com.digitalhouse.clinicaOdontologica.service.impl;
 
 
 import com.digitalhouse.clinicaOdontologica.entity.Paciente;
+import com.digitalhouse.clinicaOdontologica.exception.ResourceNotFoundException;
 import com.digitalhouse.clinicaOdontologica.repository.IPacienteRepository;
 import com.digitalhouse.clinicaOdontologica.service.IPacienteService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 
@@ -13,6 +16,8 @@ import java.util.Optional;
 @Service
 public class PacienteService implements IPacienteService {
 
+    private final Logger LOGGER = LoggerFactory.getLogger(Paciente.class);
+
     private IPacienteRepository pacienteRepository;
 
     public PacienteService(IPacienteRepository pacienteRepository) {
@@ -21,26 +26,54 @@ public class PacienteService implements IPacienteService {
 
     @Override
     public Paciente savePaciente(Paciente paciente) {
-        return pacienteRepository.save(paciente);
+        Paciente pacienteDB= null;
+        pacienteDB=pacienteRepository.save(paciente);
+        if(pacienteDB == null){
+            LOGGER.info("No se guardo ningun paciente");
+            throw new ResourceNotFoundException("El paciente no fue guardado");
+        }else{
+            LOGGER.info("Paciente  guardado " + pacienteDB.toString());
+        }
+        return pacienteDB;
     }
 
     @Override
     public Optional<Paciente> getPacienteById(Integer id) {
-        return pacienteRepository.findById(id);
+        Optional<Paciente> paciente = pacienteRepository.findById(id);
+        LOGGER.info("Paciente encontrado " + paciente.get().toString() );
+        return paciente;
     }
 
     @Override
     public List<Paciente> getAll() {
-        return pacienteRepository.findAll();
+        List<Paciente> pacientes = pacienteRepository.findAll();
+        LOGGER.info("Pacientes encontrados " + pacientes);
+        return pacientes ;
     }
 
     @Override
     public void updatePaciente(Paciente paciente) {
-        pacienteRepository.save(paciente);
+        Optional<Paciente>pacienteFound= pacienteRepository.findById(paciente.getId());
+        if(pacienteFound.isPresent()){
+            LOGGER.info("Paciente fue actualizado");
+            pacienteRepository.save(paciente);
+
+        }else{
+            LOGGER.info("El paciente "+ paciente.getId() +" no fue encontrado");
+            throw new ResourceNotFoundException("El paciente "+ paciente.getId() +" no fue encontrado");
+        }
+
     }
 
     @Override
     public void deletePacienteById(Integer id) {
-        pacienteRepository.deleteById(id);
+        Optional<Paciente>pacienteFound= pacienteRepository.findById(id);
+        if(pacienteFound.isPresent()){
+            LOGGER.info("Paciente fue eliminado");
+            pacienteRepository.deleteById(id);
+        }else{
+            LOGGER.info("El paciente "+ id +" no fue encontrado");
+            throw new ResourceNotFoundException("El paciente "+ id +" no fue encontrado");
+        }
     }
 }
